+++ NEWS.md
@@ -7,6 +7,8 @@ admiral extension packages (#1079)
 
 ## Updates of Existing Functions
 
+- `list_tte_source_objects()` gains a `package` parameter and is now exported (#1212)
+
 - `list_all_templates()` and `use_ad_templates()` gain a `package` parameter which
 can be used to indicate in which package to look for templates (#1205)